fix: move app-tab-nav feature to shared module, resolve navigation issue
