+++ src/app/app.component.html
@@ -1,9 +1,9 @@
 <ion-app>
-  <ion-split-pane contentId="main"> 
+  <ion-split-pane contentId="main">
     <ion-menu side="start" contentId="main" type="overlay">
       <app-menu></app-menu>
-    </ion-menu>  
-    <ion-router-outlet id="main"></ion-router-outlet>
+    </ion-menu>
+    <ion-router-outlet id="main">
+    </ion-router-outlet>
   </ion-split-pane>
-  <app-footer-nav> </app-footer-nav>
 </ion-app>
