--- src/app/components/footer-nav/footer-nav.page.spec.ts
@@ -1,24 +1,0 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { IonicModule } from '@ionic/angular';
-
-import { FooterNavPage } from './footer-nav.page';
-
-describe('FooterNavPage', () => {
-  let component: FooterNavPage;
-  let fixture: ComponentFixture<FooterNavPage>;
-
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
-      declarations: [ FooterNavPage ],
-      imports: [IonicModule.forRoot()]
-    }).compileComponents();
-
-    fixture = TestBed.createComponent(FooterNavPage);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  }));
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
