+++ src/app/pages/about-us/about-us.page.ts
@@ -7,7 +7,7 @@ import {
 } from 'src/app/constants/core-api';
 import { UserType } from 'src/app/models/core-api';
 import { UserThemeColorPrimary } from 'src/app/models/ui';
-import { AuthService } from 'src/app/services/auth/auth.service';
+import { AuthService } from 'src/app/shared/services/auth/auth.service';
 
 @Component({
   selector: 'app-about-us',
