+++ src/app/pages/app-settings/app-settings.module.ts
@@ -5,7 +5,7 @@ import { FormsModule } from '@angular/forms';
 import { IonicModule } from '@ionic/angular';
 
 import { AppSettingsPageRoutingModule } from './app-settings-routing.module';
-
+import { SharedModule } from 'src/app/shared/shared.module';
 import { AppSettingsPage } from './app-settings.page';
 
 @NgModule({
@@ -13,6 +13,7 @@ import { AppSettingsPage } from './app-settings.page';
     CommonModule,
     FormsModule,
     IonicModule,
+    SharedModule,
     AppSettingsPageRoutingModule,
   ],
   declarations: [AppSettingsPage],
