+++ src/app/pages/app-settings/app-settings.page.html
@@ -30,3 +30,4 @@
     </ion-label>
   </ion-item>
 </ion-content>
+<app-footer-nav currentTab="/settings" ></app-footer-nav>
