+++ src/app/pages/my-requests/my-requests.module.ts
@@ -5,6 +5,7 @@ import { FormsModule } from '@angular/forms';
 import { IonicModule } from '@ionic/angular';
 
 import { MyRequestsPageRoutingModule } from './my-requests-routing.module';
+import { SharedModule } from 'src/app/shared/shared.module';
 
 import { MyRequestsPage } from './my-requests.page';
 import { RequestComponent } from 'src/app/components/request/request.component';
@@ -14,6 +15,7 @@ import { RequestComponent } from 'src/app/components/request/request.component';
     CommonModule,
     FormsModule,
     IonicModule,
+    SharedModule,
     MyRequestsPageRoutingModule,
   ],
   declarations: [MyRequestsPage, RequestComponent],
