+++ src/app/pages/my-requests/my-requests.page.html
@@ -32,3 +32,4 @@
     </app-request>
   </ng-container>
 </ion-content>
+<app-footer-nav currentTab="/my-requests" ></app-footer-nav>
