+++ src/app/pages/my-requests/my-requests.page.ts
@@ -2,9 +2,9 @@ import { Component, OnInit, OnDestroy } from '@angular/core';
 import { Router, NavigationExtras } from '@angular/router';
 import { Subscription } from 'rxjs';
 
-import { MiscService } from 'src/app/services/misc/misc.service';
-import { CoreAPIService } from 'src/app/services/core-api/core-api.service';
-import { AuthService } from 'src/app/services/auth/auth.service';
+import { MiscService } from 'src/app/shared/services/misc/misc.service';
+import { CoreAPIService } from 'src/app/shared/services/core-api/core-api.service';
+import { AuthService } from 'src/app/shared/services/auth/auth.service';
 import { UserType } from 'src/app/models/core-api';
 import {
   defaultUserType,
