+++ src/app/pages/quarantine-map/quarantine-map.html
@@ -37,3 +37,4 @@
     </ion-fab-button>
   </ion-fab>
 </ion-content>
+<app-footer-nav currentTab="/map" ></app-footer-nav>
