+++ src/app/shared/services/core-api/core-api.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { environment } from '../../../environments/environment';
+import { environment } from '../../../../environments/environment';
 import { CommonHTTPService } from '../common-http/common-http.service';
 
 import { LatLng } from 'src/app/models/geo';
